Rationalize fish_wcstoi/d and friends
Historically fish has used the functions `fish_wcstol`, `fish_wcstoi`, and
`fish_wcstoul` (and some long long variants) for most integer conversions.
These have semantics that are deliberately different from the libc
functions, such as consuming trailing whitespace, and disallowing `-` in
unsigned versions.

fish has started to drift away from these semantics; some divergence from
C++ has crept in.

Rename the existing `fish_wcs*` functions in Rust to remove the fish
prefix, to express that they attempt to mirror libc semantics; then
introduce `fish_` wrappers which are ported from C++. Also fix some
miscellaneous bugs which have crept in, such as missing range checks.

+++ src/parser.cpp
@@ -501,7 +501,7 @@ job_t *parser_t::job_get_from_pid(pid_t pid) const {
     return job_get_from_pid(pid, job_pos);
 }
 
-job_t *parser_t::job_get_from_pid(int64_t pid, size_t &job_pos) const {
+job_t *parser_t::job_get_from_pid(int pid, size_t &job_pos) const {
     for (auto it = job_list.begin(); it != job_list.end(); ++it) {
         for (const process_ptr_t &p : (*it)->processes) {
             if (p->pid == pid) {
